Simplify release commands and publishing flow
